+++ NEWS.md
@@ -1,11 +1,15 @@
 # TreeSearch 1.6.1.9004 (development)
+- `PresCont()` implements the Group Present / Contradicted measure of
+  Goloboff et al. (2003).
 - Support single-character matrices in `ClusteringConcordance()`
 - Fix `DoNothing(x)` to return `x` (not `NULL`)
 - `Consistency()` also returns the relative homoplasy index of Steell et al. 
   (2023).
 - Remove unused `delete_rawdata()` due to implementation issues.
 
+
 # TreeSearch 1.6.1 (2025-06-10)
+
 - Handle invariant characters in `PolEscapa()`
 - Handle challenging root positions in `PlotCharacter()`
 - Fix character state colours in app legend